requirements.txt
aiohttp==3.9.0b1
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
billiard==4.2.0
boto3==1.28.85
botocore==1.31.85
celery==5.3.5
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.5
dataclasses-json==0.6.2
defusedxml==0.8.0rc2
distlib==0.3.7
distro==1.8.0
Django==5.0b1
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-filter==23.3
django-ses==3.5.0
django-storages==1.14.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
drf-nested-routers==0.93.4
drf-yasg==1.21.7
exceptiongroup==1.1.3
faiss-cpu==1.7.4
filelock==3.13.1
frozenlist==1.4.0
google-search-results==2.4.2
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
kombu==5.3.3
langchain==0.0.335
langsmith==0.0.64
lxml==4.9.3
marshmallow==3.20.1
model-bakery==1.17.0
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.0
numexpr==2.8.6
numpy==1.26.2
oauthlib==3.2.2
openai==1.2.4
packaging==23.2
Pillow==10.1.0
pipenv==2023.11.15
platformdirs==4.0.0
pluggy==1.3.0
prompt-toolkit==3.0.41
pycparser==2.21
pydantic==2.5.0
pydantic_core==2.14.1
PyJWT==2.8.0
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
route53==1.0.1
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
tenacity==8.2.3
tiktoken==0.5.1
tomli==2.0.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vine==5.1.0
virtualenv==20.24.7
wcwidth==0.2.10
yarl==1.9.2
